requirements.txt
django-social-auth
arrow
flup
python-memcached
simplejson